# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ApplyBlitz API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  targetRoles: string;
  targetLocations: string;
  yearsExperience: number;
  skills: string;
  resumeText: string;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /** @nullable */
  emailUser?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export interface ProfileInput {
  name: string;
  targetRoles: string;
  targetLocations: string;
  yearsExperience: number;
  skills: string;
  resumeText: string;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /** @nullable */
  emailUser?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
}

export interface JobListing {
  externalId: string;
  title: string;
  company: string;
  location: string;
  source: string;
  applyUrl: string;
  description: string;
  /** @nullable */
  salary?: string | null;
  /** @nullable */
  matchScore?: number | null;
  isRemote?: boolean;
  postedAt: string;
}

export interface SavedJob {
  id: number;
  externalId: string;
  title: string;
  company: string;
  location: string;
  source: string;
  applyUrl: string;
  description: string;
  /** @nullable */
  salary?: string | null;
  /** @nullable */
  matchScore?: number | null;
  isRemote?: boolean;
  postedAt: string;
  createdAt: string;
}

export interface SavedJobInput {
  externalId: string;
  title: string;
  company: string;
  location: string;
  source: string;
  applyUrl: string;
  description: string;
  /** @nullable */
  salary?: string | null;
  /** @nullable */
  matchScore?: number | null;
  isRemote?: boolean;
  postedAt: string;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  applied: 'applied',
  interviewing: 'interviewing',
  rejected: 'rejected',
  offer: 'offer',
  ghosted: 'ghosted',
} as const;

export interface Application {
  id: number;
  company: string;
  role: string;
  status: ApplicationStatus;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tailoredResume?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  matchScore?: number | null;
  appliedAt: string;
  createdAt: string;
  updatedAt?: string;
}

export type ApplicationInputStatus = typeof ApplicationInputStatus[keyof typeof ApplicationInputStatus];


export const ApplicationInputStatus = {
  applied: 'applied',
  interviewing: 'interviewing',
  rejected: 'rejected',
  offer: 'offer',
  ghosted: 'ghosted',
} as const;

export interface ApplicationInput {
  company: string;
  role: string;
  status?: ApplicationInputStatus;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tailoredResume?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  matchScore?: number | null;
  /** @nullable */
  appliedAt?: string | null;
}

export type ApplicationUpdateStatus = typeof ApplicationUpdateStatus[keyof typeof ApplicationUpdateStatus];


export const ApplicationUpdateStatus = {
  applied: 'applied',
  interviewing: 'interviewing',
  rejected: 'rejected',
  offer: 'offer',
  ghosted: 'ghosted',
} as const;

export interface ApplicationUpdate {
  status?: ApplicationUpdateStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tailoredResume?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  matchScore?: number | null;
}

export interface BatchApplicationInput {
  jobs: ApplicationInput[];
  dryRun?: boolean;
}

export interface BatchApplicationResult {
  created: number;
  dryRun: boolean;
  applications?: Application[];
}

export interface TailorInput {
  jobTitle: string;
  company: string;
  jobDescription: string;
  /** @nullable */
  resumeText?: string | null;
}

export interface TailorResult {
  tailoredResume: string;
  coverLetter: string;
  matchScore: number;
  keywords: string[];
}

export interface MatchScoreInput {
  resumeText: string;
  jobDescription: string;
}

export interface MatchScoreResult {
  score: number;
  matchedKeywords: string[];
  missingKeywords: string[];
}

export type DashboardStatsStatusBreakdown = {
  applied: number;
  interviewing: number;
  rejected: number;
  offer: number;
  ghosted: number;
};

export interface DashboardStats {
  totalApplications: number;
  appliedToday: number;
  dailyGoal: number;
  streakDays: number;
  responseRate: number;
  weeklyCount: number;
  statusBreakdown: DashboardStatsStatusBreakdown;
}

export interface DailyStatEntry {
  date: string;
  count: number;
}

export type SearchJobsParams = {
query: string;
location?: string;
remote?: boolean;
/**
 * @nullable
 */
minSalary?: number | null;
/**
 * @nullable
 */
maxSalary?: number | null;
datePosted?: SearchJobsDatePosted;
source?: string;
};

export type SearchJobsDatePosted = typeof SearchJobsDatePosted[keyof typeof SearchJobsDatePosted];


export const SearchJobsDatePosted = {
  today: 'today',
  week: 'week',
  month: 'month',
} as const;

export type GetApplicationsParams = {
status?: GetApplicationsStatus;
search?: string;
};

export type GetApplicationsStatus = typeof GetApplicationsStatus[keyof typeof GetApplicationsStatus];


export const GetApplicationsStatus = {
  applied: 'applied',
  interviewing: 'interviewing',
  rejected: 'rejected',
  offer: 'offer',
  ghosted: 'ghosted',
} as const;