# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ApplyBlitz API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: getApplications
      tags: [applications]
      summary: List all applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [applied, interviewing, rejected, offer, ghosted]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Log a new job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a single application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /applications/batch:
    post:
      operationId: batchCreateApplications
      tags: [applications]
      summary: Batch apply to multiple jobs at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchApplicationInput"
      responses:
        "201":
          description: Batch apply result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchApplicationResult"

  # === AI TAILORING ===
  /ai/tailor:
    post:
      operationId: tailorResume
      tags: [ai]
      summary: AI-tailor resume and generate cover letter for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TailorInput"
      responses:
        "200":
          description: Tailored resume + cover letter + match score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TailorResult"
  /ai/match-score:
    post:
      operationId: getMatchScore
      tags: [ai]
      summary: Get ATS keyword match score between resume and job description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchScoreInput"
      responses:
        "200":
          description: Match score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchScoreResult"

  # === STATS ===
...
components:
  schemas:
...
    Application:
      type: object
      required: [id, company, role, status, appliedAt, createdAt]
      properties:
        id:
          type: integer
        company:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [applied, interviewing, rejected, offer, ghosted]
        applyUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tailoredResume:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        matchScore:
          type: ["integer", "null"]
        appliedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    ApplicationInput:
      type: object
      required: [company, role]
      properties:
        company:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [applied, interviewing, rejected, offer, ghosted]
        applyUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tailoredResume:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        matchScore:
          type: ["integer", "null"]
        appliedAt:
          type: ["string", "null"]
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [applied, interviewing, rejected, offer, ghosted]
        notes:
          type: ["string", "null"]
        tailoredResume:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        matchScore:
          type: ["integer", "null"]
    BatchApplicationInput:
      type: object
      required: [jobs]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationInput"
        dryRun:
          type: boolean
    BatchApplicationResult:
      type: object
      required: [created, dryRun]
      properties:
        created:
          type: integer
        dryRun:
          type: boolean
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"

    # --- AI ---
    TailorInput:
      type: object
      required: [jobTitle, company, jobDescription]
      properties:
        jobTitle:
          type: string
        company:
          type: string
        jobDescription:
          type: string
        resumeText:
          type: ["string", "null"]
    TailorResult:
      type: object
      required: [tailoredResume, coverLetter, matchScore, keywords]
      properties:
        tailoredResume:
          type: string
        coverLetter:
          type: string
        matchScore:
          type: integer
        keywords:
          type: array
          items:
            type: string
    MatchScoreInput:
      type: object
      required: [resumeText, jobDescription]
      properties:
        resumeText:
          type: string
        jobDescription:
          type: string
    MatchScoreResult:
      type: object
      required: [score, matchedKeywords, missingKeywords]
      properties:
        score:
          type: integer
        matchedKeywords:
          type: array
          items:
            type: string
        missingKeywords:
          type: array
          items:
            type: string

    # --- Stats ---
...