# lib/api-zod/src/generated/types/batchApplicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ApplyBlitz API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationInput } from './applicationInput';

export interface BatchApplicationInput {
  jobs: ApplicationInput[];
  dryRun?: boolean;
}