# lib/api-zod/src/generated/types/batchApplicationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ApplyBlitz API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';

export interface BatchApplicationResult {
  created: number;
  dryRun: boolean;
  applications?: Application[];
}