# lib/api-zod/src/generated/types/tailorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ApplyBlitz API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TailorInput {
  jobTitle: string;
  company: string;
  jobDescription: string;
  /** @nullable */
  resumeText?: string | null;
}